Add Product model, migrations, factory, and seed data
This commit introduces several changes related to addition of 'Product' functionality. A new 'Products' table is created in the database through migration with fields like title, quantity, price, description, category, image, ratings_rate, ratings_count, etc. An Eloquent model 'Product' is also added which uses Laravel's mass assignment for filling specific data in the database, and treats ratings as an array. Further, a 'ProductFactory' has been included for generating test instances of 'Product', and 'Product' seed data is added to the DatabaseSeeder. This setup will facilitate managing products data in the application.

Additionally, 'throttle' value in 'password_resets' of auth config is updated to 600 for increasing the limit, a route to fetch 'Product' data is added, and password reset throttle is adjusted for user convenience in account recovery situation. This update ensures a smoother user experience with increased security.